requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
opencv-python
pillow
seaborn
streamlit
